feat: move CalTopo features to the CalTopo context menu
This change moves the CalTopo features (e.g. Open in Mountain Project)
from the Chrome context menu to the CalTopo context menu.  CalTopo
disables the Chrome context menu by default.  Previous to this change,
an override of this default was required.

Because Mountain Project requires a script to be injected into the page
in order to jump to coordinates, this change modifies how the CalTopo
page script opens new URLs.  Previous to this change, the URLs were
opened using the window.open API.  To support script injection for
Mountain Project, URLs are now opened by passing a message to the
extension.  The extension then opens the URL and injects any necessary
scripts.