fix: support for Google Maps satellite view
Previous to this change, the Google Maps context menu options did not
work when in satellite view.  This was due to satellite view using a
different URL format.  It uses an elevation parameter instead of a zoom
parameter.

After collecting the elevation equivalent for each zoom level.  I
plotted the data in Excel and added a logarithmic trendline to find the
equation to convert elevation to zoom.